+++ nipype/workflows/smri/freesurfer/__init__.py
@@ -1,3 +1,3 @@
 from .utils import (create_getmask_flow, create_get_stats_flow, create_tessellation_flow)
 from .bem import create_bem_flow
-from .recon import create_skullstripped_recon_flow
+from .recon import create_skullstripped_recon_flow, create_reconall_workflow
