+++ nipype/workflows/smri/freesurfer/autorecon1.py
@@ -7,8 +7,7 @@ from nipype.interfaces.freesurfer import *
 from utils import copy_file, copy_files, awkfile
 
 def checkT1s(T1_files, cw256=False):
-    """Verify size outside of pipeline processing"""
-    print("Verifying input T1 size")
+    """Verifying size of inputs and setting workflow parameters"""
     import SimpleITK as sitk
     import os
     import sys
@@ -24,7 +23,14 @@ def checkT1s(T1_files, cw256=False):
             convert.run()
             T1_files[i] = convert.inputs.out_file
     size = None
-    for t1 in T1_files:
+    origvol_names = list()
+    for i, t1 in enumerate(T1_files):
+        # assign an input number
+        file_num = str(i + 1)
+        while len(file_num) < 3:
+            file_num = '0' + file_num
+        origvol_names.append("{0}.mgz".format(file_num))
+        # check the size of the image
         img = sitk.ReadImage(t1)
         if not size:
             size = img.GetSize()
@@ -32,31 +38,20 @@ def checkT1s(T1_files, cw256=False):
             print("ERROR: T1s not the same size. Cannot process {0} {1} together".format(T1_files[0],
                                                                                          otherFilename))
             sys.exit(-1)
-    # check if cw256 is set to crop the images
+    # check if cw256 is set to crop the images if size is larger than 256
     if not cw256:
         for dim in size:
             if dim > 256:
                 print("Setting MRI Convert to crop images to 256 FOV")
                 cw256 = True
-    origvols, iscaleout, ltaout = create_preproc_filenames(T1_files)
-    return T1_files, cw256, origvol_names, 
+    if len(T1_files) > 1:
+        resample_type = 'cubic'
+    else:
+        resample_type = 'interpolate'
+    return T1_files, cw256, resample_type, origvol_names
 
-def create_preproc_filenames(in_T1s):
-    # Create output filenames
-    inputvols = list()
-    iscaleout = list()
-    ltaout = list()
-
-    for i, T1 in enumerate(in_T1s):
-        file_num = str(i + 1)
-        while len(file_num) < 3:
-            file_num = '0' + file_num
-        iscaleout.append(file_num + '-iscale.txt')
-        ltaout.append(file_num + '.lta')
-        inputvols.append(file_num + '.mgz')
-    return inputvols, iscaleout, ltaout
-
-def create_AutoRecon1(name="AutoRecon1", longitudinal=False, use_T2=False, use_FLAIR=False, config):
+def create_AutoRecon1(name="AutoRecon1", longitudinal=False, field_strength='1.5T',
+                      custom_atlas=None, awk_file=None, plugin_args=None):
     """Creates the AutoRecon1 workflow in nipype.
 
     Inputs::
@@ -64,26 +59,28 @@ def create_AutoRecon1(name="AutoRecon1", longitudinal=False, use_T2=False, use_F
            inputspec.T2_file : T2 file (optional)
            inputspec.FLAIR_file : FLAIR file (optional)
            inputspec.cw256 : Conform inputs to 256 FOV (optional)
-
+           inputspec.num_threads: Number of threads to use with EM Register (default=1)
     Outpus::
            
     """
     ar1_wf = pe.Workflow(name=name)
+    inputspec = pe.Node(interface=IdentityInterface(fields=['T1_files',
+                                                            'T2_file',
+                                                            'in_flair',
+                                                            'cw256',
+                                                            'num_threads',
+                                                            'skulltemplate']),
+                        run_without_submitting=True,
+                        name='inputspec')
 
     if not longitudinal:
         # single session processing
-        inputspec = pe.Node(interface=IdentityInterface(
-            fields=['T1_files', 'T2_file', 'in_flair', 'cw256']),
-                             run_without_submitting=True,
-                             name='inputspec')
-
         verify_inputs = pe.Node(Function(infields=["T1_files", "cw256"],
-                                         outfields=["T1_files", "cw256"],
+                                         outfields=["T1_files", "cw256", "resample_type", "origvol_names"],
                                          checkT1s)
                                 name="Check_T1s"),
         ar1_wf.conncet([(inputspec, verify_inputs, [('T1_files', 'T1_files'),
                                                     ('cw256', 'cw256')])])
-
 
 
         # T1 image preparation
@@ -91,61 +88,82 @@ def create_AutoRecon1(name="AutoRecon1", longitudinal=False, use_T2=False, use_F
         T1_image_preparation = pe.MapNode(MRIConvert(),
                                           iterfield=['in_file', 'out_file'],
                                           name="T1_prep")
-        T1_image_preparation.inputs.out_file = origvols
 
-        ar1_wf.connect([(inputspec, T1_image_preparation, [('T1_files', 'in_file')]),
+        ar1_wf.connect([(verify_inputs, T1_image_preparation, [('T1_files', 'in_file'),
+                                                               ('origvol_names', 'out_file')]),
                         ])
 
-        # T2 image preparation
-        if use_T2:
-            # Create T2raw.mgz
-            # mri_convert
-            inputspec.inputs.T2_file = config['in_T2']
-            T2_convert = pe.Node(MRIConvert(), name="T2_convert")
-            T2_convert.inputs.out_file = 'T2raw.mgz'
-            T2_convert.inputs.no_scale = True
-            ar1_wf.connect([(inputspec, T2_convert, [('T2_file', 'in_file')]),
-                            ]) 
+        def convert_modalities(in_file, out_file):
+            """Returns an undefined output if the in_file is not defined"""
+            from nipype.interfaces.base import isdefined
+            from nipype.interfaces.freesurfer import MRIConvert
+            if isdefined(in_file):
+                convert = MRIConvert()
+                convert.inputs.in_file = in_file
+                convert.inputs.out_file = out_file
+                convert.inputs.no_scale = True
+                convert.run()
+                out_file = os.path.abspath(convert.outputs.out_file)
+            return out_file
 
-        # FLAIR image preparation
-        if use_FLAIR:
-            # Create FLAIRraw.mgz
-            # mri_convert
-            inputspec.inputs.in_flair = config['in_FLAIR']
-            FLAIR_convert = pe.Node(MRIConvert(), name="FLAIR_convert")
-            FLAIR_convert.inputs.out_file = 'FLAIRraw.mgz'
-            FLAIR_convert.inputs.no_scale = True
-            ar1_wf.connect([(inputspec, FLAIR_convert, [('in_flair', 'in_file')]),
-                            ])
+        T2_convert = pe.Node(Function(['in_file', 'out_file'],
+                                      ['out_file'],
+                                      convert_modalities),
+                             name="T2_Convert")
+        T2_convert.inputs.out_file = 'T2raw.mgz'
+        ar1_wf.connect([(inputspec, T2_convert, [('T2_file', 'in_file')])]) 
+
+        FLAIR_convert = pe.Node(Function(['in_file', 'out_file'],
+                                         ['out_file'],
+                                         convert_modalities),
+                                name="T2_Convert")
+        FLAIR_convert.inputs.out_file = 'FLAIRraw.mgz'
+        ar1_wf.connect([(inputspec, FLAIR_convert, [('in_flair', 'in_file')])])        
     else:
         # longitudinal inputs
-        inputspec = pe.Node(interface=IdentityInterface(
-            fields=['T1_files',
-                    'iscales',
-                    'ltas',
-                    'subj_to_template_lta',
-                    'template_talairach_xfm',
-                    'template_brainmask']),
-                             run_without_submitting=True,
-                             name='inputspec')
+        inputspec = pe.Node(interface=IdentityInterface(fields=['T1_files',
+                                                                'iscales',
+                                                                'ltas',
+                                                                'subj_to_template_lta',
+                                                                'template_talairach_xfm',
+                                                                'template_brainmask']),
+                            run_without_submitting=True,
+                            name='inputspec')
 
-        _volnames_, in_iscales, in_ltas = create_preproc_filenames(config['in_T1s'])
+        def output_names(T1_files):
+            """Create file names that are dependent on the number of T1 inputs"""
+            iscale_names = list()
+            lta_names = list()
+            for i, t1 in enumerate(T1_files):
+                # assign an input number
+                file_num = str(i + 1)
+                while len(file_num) < 3:
+                    file_num = '0' + file_num
+                iscale_names.append("{0}-iscale.txt".format(file_num))
+                lta_names.append("{0}.lta".format(file_num))
+            return iscale_names, lta_names
 
+        filenames = pe.Node(Function(['T1_files'],
+                                     ['iscale_names', 'lta_names']
+                                     output_names),
+                            name="Longitudinal_Filenames")
+        ar1_wf.connect([(inputspec, filenames, [('T1_files', 'T1_files')])])
+        
         copy_ltas = pe.MapNode(Function(['in_file', 'out_file'],
                                         ['out_file'],
                                         copy_file),
                                iterfield=['in_file', 'out_file'],
                                name='Copy_ltas')
-        ar1_wf.connect([(inputspec, copy_ltas, [('ltas', 'in_file')])])
-        copy_ltas.inputs.out_file = in_ltas
+        ar1_wf.connect([(inputspec, copy_ltas, [('ltas', 'in_file')]),
+                        (filenames, copy_ltas, [('lta_names', 'out_file')])])
 
         copy_iscales = pe.MapNode(Function(['in_file', 'out_file'],
                                            ['out_file'],
                                            copy_file),
                                   iterfield=['in_file', 'out_file'],
                                   name='Copy_iscales')
-        ar1_wf.connect([(inputspec, copy_iscales, [('iscales', 'in_file')])])
-        copy_iscales.inputs.out_file = in_iscales
+        ar1_wf.connect([(inputspec, copy_iscales, [('iscales', 'in_file')]),
+                        (filenames, copy_iscales, [('iscale_names', 'out_file')])])
 
         concatenate_lta = pe.MapNode(ConcatenateLTA(), iterfield=['in_file'],
                                      name="Concatenate_ltas")
@@ -161,59 +179,64 @@ def create_AutoRecon1(name="AutoRecon1", longitudinal=False, use_T2=False, use_F
     255 cubed char images (1mm isotropic voxels) in mri/orig.mgz.
     """
 
-    if not config['longitudinal'] and len(config['in_T1s']) == 1:
-        # if only one T1 scan is input, just copy the converted scan as the rawavg.mgz
-        create_template = pe.Node(Function(['in_file', 'out_file'],
-                                           ['out_file'],
-                                            copy_file),
+    def createTemplate(in_files, out_file):
+        if len(in_files) == 1 and not longitudinal:
+            print("WARNING: only one run found. This is OK, but motion correction" +
+                  "cannot be performed on one run, so I'll copy the run to rawavg" +
+                  "and continue.")
+            copy_file(in_files[0], out_file)
+            intensity_scales = None
+            transforms = None
+        else:
+            from nipype.interfaces.freesurfer import RobustTemplate
+            # if multiple T1 scans are given
+            intensity_scales = [os.path.basename(f.replace('.mgz', '-iscale.txt')) for f in in_files]
+            transforms = [os.path.basename(f.replace('.mgz', '.lta')) for f in in_files]
+            robtemp = RobustTemplate()
+            robtemp.inputs.average_metric = 'median'
+            robtemp.inputs.out_file = out_file
+            robtemp.inputs.no_iteration = True
+            robtemp.inputs.fixed_timepoint = True
+            robtemp.inputs.auto_detect_sensitivity = True
+            robtemp.inputs.initial_timepoint = 1
+            robtemp.inputs.scaled_intensity_outputs = intensity_scales
+            robtemp.inputs.transform_outputs = transforms
+            robtemp.inputs.subsample_threshold = 200
+            robtemp.inputs.intensity_scaling = True
+            robtemp.run()
+            intensity_scales = [os.path.abspath(f) for f in robtemp.outputs.scaled_intensity_outputs]
+            transforms = [os.path.abspath(f) for f in robtemp.outputs.transform_outputs]
+            out_file = robtemp.outputs.out_file
+        out_file = os.path.abspath(out_file)
+        return out_file, intensity_scales, transforms
+    
+    if not longitudinal:
+        create_template = pe.Node(Function(['in_files', 'out_file'],
+                                           ['out_file', 'intensity_scales', 'transforms'],
+                                           createTemplate),
                                   name="Robust_Template")
         create_template.inputs.out_file = 'rawavg.mgz'
-        
-        ar1_wf.connect([(T1_image_preparation, create_template, [('out_file', 'in_file')]),
-                    ])
-        
-        print("WARNING: only one run found. This is OK, but motion correction" +
-              "cannot be performed on one run, so I'll copy the run to rawavg" +
-              "and continue.")
-        
+        ar1_wf.connect([(T1_image_preparation, create_template, [('out_file', 'in_files')])])
     else:
-        # if multiple T1 scans are given
         create_template = pe.Node(RobustTemplate(), name="Robust_Template")
         create_template.inputs.average_metric = 'median'
         create_template.inputs.out_file = 'rawavg.mgz'
         create_template.inputs.no_iteration = True
-            
-        if config['longitudinal']:
-            # if running longitudinally
-            ar1_wf.connect([(concatenate_lta, create_template, [('out_file', 'initial_transforms')]),
-                            (inputspec, create_template, [('T1_files', 'in_files')]),
-                            (copy_iscales, create_template, [('out_file','in_intensity_scales')]),
-                        ])
-        else:
-            # if running single session
-            create_template.inputs.fixed_timepoint = True
-            create_template.inputs.auto_detect_sensitivity = True
-            create_template.inputs.initial_timepoint = 1
-            create_template.inputs.scaled_intensity_outputs = iscaleout
-            create_template.inputs.transform_outputs = ltaout
-            create_template.inputs.subsample_threshold = 200
-            create_template.inputs.intensity_scaling = True
-            ar1_wf.connect([(T1_image_preparation, create_template, [('out_file', 'in_files')]),
-                        ])
+        ar1_wf.connect([(concatenate_lta, create_template, [('out_file', 'initial_transforms')]),
+                        (inputSpec, create_template, [('in_t1s', 'in_files')]),
+                        (copy_iscales, create_template, [('out_file','in_intensity_scales')])])
 
     # mri_convert
     conform_template = pe.Node(MRIConvert(), name='Conform_Template')
     conform_template.inputs.out_file = 'orig.mgz'
-    if config['longitudinal']:
-        conform_template.inputs.out_datatype = 'uchar'
-    else:
+    if not longitudinal:
         conform_template.inputs.conform = True
-        if len(config['in_T1s']) != 1:
-            conform_template.inputs.cw256 = config['cw256']    
-            conform_template.inputs.resample_type = 'cubic'
+        ar1_wf.connect([(verify_inputs, conform_template, [('cw256', 'cw256'),
+                                                           ('resample_type', 'resample_type')])]) 
+    else:
+        conform_template.inputs.out_datatype = 'uchar'
             
-    ar1_wf.connect(
-        [(create_template, conform_template, [('out_file', 'in_file')])])
+    ar1_wf.connect([(create_template, conform_template, [('out_file', 'in_file')])])
 
     # Talairach
     """
@@ -247,7 +270,15 @@ def create_AutoRecon1(name="AutoRecon1", longitudinal=False, use_T2=False, use_F
     ar1_wf.connect([(conform_template, bias_correction, [('out_file', 'in_file')]),
                 ])
 
-    if config['longitudinal']:
+    if not longitudinal:
+        # single session processing
+        talairach_avi = pe.Node(TalairachAVI(), name="Compute_Transform")
+        if custom_atlas != None:
+            # allows to specify a custom atlas
+            talairach_avi.inputs.atlas = custom_atlas
+        talairach_avi.inputs.out_file = 'talairach.auto.xfm'
+        ar1_wf.connect([(bias_correction, talairach_avi, [('out_file', 'in_file')])])
+    else:
         # longitudinal processing
         # Just copy the template xfm
         talairach_avi = pe.Node(Function(['in_file', 'out_file'],
@@ -257,18 +288,6 @@ def create_AutoRecon1(name="AutoRecon1", longitudinal=False, use_T2=False, use_F
         talairach_avi.inputs.out_file = 'talairach.auto.xfm'
 
         ar1_wf.connect([(inputspec, talairach_avi, [('template_talairach_xfm', 'in_file')])])
-    else:
-        # single session processing
-        talairach_avi = pe.Node(TalairachAVI(), name="Compute_Transform")
-        
-        if config['custom_atlas'] != None:
-            # allows to specify a custom atlas
-            talairach_avi.inputs.atlas = config['custom_atlas']
-            
-        talairach_avi.inputs.out_file = 'talairach.auto.xfm'
-        
-        ar1_wf.connect([(bias_correction, talairach_avi, [('out_file', 'in_file')]),
-                    ])
         
     copy_transform = pe.Node(Function(['in_file', 'out_file'],
                                       ['out_file'],
@@ -312,19 +331,31 @@ def create_AutoRecon1(name="AutoRecon1", longitudinal=False, use_T2=False, use_F
     ar1_wf.connect([(copy_transform, check_alignment, [('out_file', 'in_file')]),
                     ])
 
-    if not config['longitudinal']:
-        awk_logfile = pe.Node(Function(['awk_file', 'log_file'],
+    if not longitudinal and awk_file:
+        def awkfile(in_file, log_file):
+            """
+            This method uses 'awk' which must be installed prior to running the workflow and is not a
+            part of nipype or freesurfer.
+            """
+            import subprocess
+            import os
+            command = ['awk', '-f', awk_file, log_file]
+            print(''.join(command))
+            subprocess.call(command)
+            log_file = os.path.abspath(log_file)
+            return log_file
+
+        awk_logfile = pe.Node(Function(['in_file', 'log_file'],
                                        ['log_file'],
                                        awkfile),
                               name='Awk')
-        awk_logfile.inputs.awk_file = config['awk_file']
+        awk_logfile.inputs.in_file = awk_file
                                        
         ar1_wf.connect([(talairach_avi, awk_logfile, [('out_log', 'log_file')])])
 
         # TODO datasink the output from TalirachQC...not sure how to do this
         tal_qc = pe.Node(TalairachQC(), name="Detect_Aligment_Failures")
-        ar1_wf.connect([(awk_logfile, tal_qc, [('log_file', 'log_file')]),
-                    ])
+        ar1_wf.connect([(awk_logfile, tal_qc, [('log_file', 'log_file')])])
 
     # Intensity Normalization
     # Performs intensity normalization of the orig volume and places the result in mri/T1.mgz.
@@ -345,12 +376,30 @@ def create_AutoRecon1(name="AutoRecon1", longitudinal=False, use_T2=False, use_F
     Removes the skull from mri/T1.mgz and stores the result in 
     mri/brainmask.auto.mgz and mri/brainmask.mgz. Runs the mri_watershed program.
     """
+    if not longitudinal:    
+        mri_em_register = pe.Node(EMRegister(), name="EM_Register")
+        mri_em_register.inputs.out_file = 'talairach_with_skull.lta'
+        mri_em_register.inputs.skull = True
+        if plugin_args:
+            mri_em_register.plugin_args = plugin_args
+            
+        ar1_wf.connect([(add_xform_to_orig_nu, mri_em_register, [('out_file', 'in_file')]),
+                        (inputspec, mri_em_register, [('num_threads', 'num_threads'),
+                                                      ('skulltemplate', 'template')])])
 
-    if config['longitudinal']:
+        brainmask = pe.Node(WatershedSkullStrip(),
+                            name='Watershed_Skull_Strip')
+        brainmask.inputs.t1 = True
+        brainmask.inputs.out_file = 'brainmask.auto.mgz'
+        ar1_wf.connect([(mri_normalize, brainmask, [('out_file', 'in_file')]),
+                        (mri_em_register, brainmask, [('out_file', 'transform')]),
+                        (inputspec, brainmask, [('skulltemplate', 'brain_atlas')])])
+    else:
         copy_template_brainmask = pe.Node(Function(['in_file', 'out_file'],
                                                    ['out_file'],
                                                    copy_file),
                                           name='Copy_Template_Brainmask')
+        #TODO: Change this to inputspec
         copy_template_brainmask.inputs.out_file = 'brainmask_{0}.mgz'.format(config['long_template'])
         
         ar1_wf.connect([(inputspec, copy_template_brainmask, [('template_brainmask', 'in_file')])])
@@ -370,30 +419,6 @@ def create_AutoRecon1(name="AutoRecon1", longitudinal=False, use_T2=False, use_F
         ar1_wf.connect([(mask1, brainmask, [('out_file', 'in_file')]),
                         (copy_template_brainmask, brainmask, [('out_file', 'mask_file')])])
                         
-    else:    
-        mri_em_register = pe.Node(EMRegister(), name="EM_Register")
-        mri_em_register.inputs.template = config['registration_template']
-        mri_em_register.inputs.out_file = 'talairach_with_skull.lta'
-        mri_em_register.inputs.skull = True
-        
-        if config['openmp'] != None:
-            mri_em_register.inputs.num_threads = config['openmp']
-        if config['plugin_args'] != None:
-            mri_em_register.plugin_args = config['plugin_args']
-            
-        ar1_wf.connect([(add_xform_to_orig_nu, mri_em_register, [('out_file', 'in_file')]),
-                    ])
-
-        brainmask = pe.Node(
-            WatershedSkullStrip(), name='Watershed_Skull_Strip')
-        brainmask.inputs.t1 = True
-        brainmask.inputs.brain_atlas = config['registration_template']
-        brainmask.inputs.out_file = 'brainmask.auto.mgz'
-        ar1_wf.connect([(mri_normalize, brainmask, [('out_file', 'in_file')]),
-                        (mri_em_register, brainmask,
-                         [('out_file', 'transform')]),
-                    ])
-
     copy_brainmask = pe.Node(Function(['in_file', 'out_file'],
                                       ['out_file'],
                                       copy_file),
@@ -432,11 +457,11 @@ def create_AutoRecon1(name="AutoRecon1", longitudinal=False, use_T2=False, use_F
                     ])
 
     
-    if config['longitudinal']:
-        ar1_wf.connect([(copy_template_brainmask, outputspec, [('out_file', 'braintemplate')]),
+    if not longitudinal:
+        ar1_wf.connect([(mri_em_register, outputspec, [('out_file', 'talskull')]),
                         ])
     else:
-        ar1_wf.connect([(mri_em_register, outputspec, [('out_file', 'talskull')]),
+        ar1_wf.connect([(copy_template_brainmask, outputspec, [('out_file', 'braintemplate')]),
                         ])
 
     return ar1_wf, outputs
